# src/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/signup": {
      "post": {
        "tags": [
          "Sign up"
        ],
        "summary": "Create a new user account",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "User",
            "description": "The user account to create",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ]
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Log into user account",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "User",
            "description": "The user account to authenticate",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ]
      }
    },
    "/request-reset": {
      "post": {
        "tags": [
          "Request a password reset"
        ],
        "summary": "Request an email with a password reset link",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "Email",
            "description": "The email account to send the reset link to",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email"
              ],
              "properties": {
                "email": {
                  "type": "string"
                }
              }
            }
          }
        ]
      }
    },
    "/reset-password": {
      "post": {
        "tags": [
          "Reset the password"
        ],
        "summary": "Use the token from the link in the email to reset the password",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "Password reset params",
            "description": "The new password and the token from the email link",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email",
                "token"
              ],
              "properties": {
                "email": {
                  "type": "string"
                },
                "token": {
                  "type": "string"
                }
              }
            }
          }
        ]
      }
    }
  }
}